backend/requirements.txt
# Core
fastapi==0.132.0
uvicorn==0.41.0

# Database
sqlalchemy==2.0.46
alembic==1.18.4

# Security & Auth
bcrypt==5.0.0
python-jose==3.5.0

# Validation & Config
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
email-validator==2.3.0
python-multipart==0.0.22
